Adding AuthForm separate component
